DLO1-53 Cleanup. Move non-script files to utilities folder Added dbConfig file to centralize DB configuration
