+++ database/python_scripts/.gitignore
@@ -6,3 +6,6 @@
 
 #python cache
 __pycache__
+
+#configuration file
+dbConfig.py
