+++ database/python_scripts/utilities/DBClear.py
